ADM-1667 Don't try to sort empty index of files to prevent 500 error
When the GUI requests the backend to sort on a column and that index is
empty we get a 500 error back that a non-existent field cannot be
sorted. Although technically an error this is a nasty result to send the
front end.

This has been resolved by doing a quick metadata only search on the
index checking for any documents and returning an empty set if the count
of documents is 0 instead of trying to sort the results.

This does add an additional search request per search request and might
no longer be required in the future if we start adding more detail to
our index documents including some mappings.