ADM-1538 Handle memcache name change and decrease unnecessary ES docs
The memcached container name has changed so it has been updated. This
also changes the caching to set a placeholder value for the id before it
creates the elastic search index. This will greatly decrease the need
for creating unnecessary docs and then deleting them when we discover
they have been created already.